Added Control Panel for FileList
